refactor(blocks-engine): move the remote-host policy into the engine
