+++ packages/blocks-engine/package.json
@@ -42,13 +42,15 @@
     "node": ">=20.0.0"
   },
   "dependencies": {
-    "css-tree": "^2.3.1"
+    "css-tree": "^2.3.1",
+    "picomatch": "^4.0.5"
   },
   "devDependencies": {
     "@nextlyhq/eslint-config": "workspace:*",
     "@nextlyhq/tsconfig": "workspace:*",
     "@types/css-tree": "^2.3.4",
     "@types/node": "^20.19.17",
+    "@types/picomatch": "^4.0.3",
     "eslint": "^9.39.1",
     "rimraf": "^6.1.3",
     "tsup": "^8.5.0",
